--- Authentication/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "1-Passport-Example",
-  "version": "1.0.0",
-  "description": "",
-  "main": "server.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server.js",
-    "watch": "nodemon server.js"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "bcryptjs": "^2.4.3",
-    "dotenv": "^8.2.0",
-    "express": "^4.17.1",
-    "express-session": "^1.17.0",
-    "fs": "0.0.1-security",
-    "morgan": "^1.9.1",
-    "mysql": "^2.17.1",
-    "mysql2": "^1.6.5",
-    "passport": "^0.4.1",
-    "passport-local": "^1.0.0",
-    "path": "^0.12.7",
-    "sequelize": "^5.21.3"
-  }
-}
