--- Authentication/public/stylesheets/style.css
@@ -1,4 +1,0 @@
-form.signup,
-form.login {
-  margin-top: 50px;
-}
