--- Authentication/routes/api-routes.js
@@ -1,77 +1,0 @@
-// Requiring our models and passport as we've configured it
-var db = require("../models");
-var passport = require("../config/passport");
-let bcrypt = require("bcryptjs");
-let logger = require("morgan");
-
-module.exports = function(app) {
-  // Using the passport.authenticate middleware with our local strategy.
-  // If the user has valid login credentials, send them to the members page.
-  // Otherwise the user will be sent an error
-  app.post("/api/login", passport.authenticate("local"), function(req, res) {
-    res.json(req.user);
-  });
-
-  // Route for signing up a user. The user's password is automatically hashed and stored securely thanks to
-  // how we configured our Sequelize User Model. If the user is created successfully, proceed to log the user in,
-  // otherwise send back an error
-  app.post("/api/signup", function(req, res) {
-    db.User.create({
-      firstName: req.body.firstName,
-      lastName: req.body.lastName,
-      email: req.body.email,
-      password: req.body.password
-    })
-      .then(function() {
-        res.redirect(307, "/api/login");
-      })
-      .catch(function(err) {
-        res.status(401).json(err);
-      });
-  });
-
-  // Route for logging user out
-  app.get("/logout", function(req, res) {
-    req.logout();
-    res.redirect("/");
-  });
-
-  // Route for getting some data about our user to be used client side
-  app.get("/api/user_data", function(req, res) {
-    if (!req.user) {
-      // The user is not logged in, send back an empty object
-      res.json({});
-    } else {
-      // Otherwise send back the user's email and id
-      // Sending back a password, even a hashed password, isn't a good idea
-      res.json({
-        email: req.user.email,
-        id: req.user.id
-      });
-    }
-  });
-
-  //this put route is still under progress
-
-  // app.put("api/updatepassword/:id", function(req, res) {
-  //   console.log(`we hit the update rtoute`);
-  //   const user_id = req.params.id;
-  //   db.findOne({
-  //     where: { id: user_id }
-  //   }).then(function(data) {
-  //     console.log(data);
-  //     bcrypt
-  //       .compare(req.body.oldPassword, data.password)
-  //       .then(function(result) {
-  //         console.log(result);
-  //       });
-  //   });
-
-  //     db.update({ password: req.body.newPassword }, { where: user_id }).then(
-  //       function(result) {
-  //         console.log(result);
-  //       }
-  //     );
-  //     res.json({ test: "true" });
-  //   });
-};
