+++ package.json
@@ -33,8 +33,17 @@
     "express": "^4.17.1",
     "if-env": "^1.0.4",
     "jimp": "^0.9.3",
-    "mongoose": "^5.8.1",
     "morgan": "^1.9.1",
-    "react-leaf-carousel": "^2.0.0"
+    "react-leaf-carousel": "^2.0.0",
+    "bcryptjs": "^2.4.3",
+    "dotenv": "^8.2.0",
+    "express-session": "^1.17.0",
+    "fs": "0.0.1-security",
+    "mysql": "^2.17.1",
+    "mysql2": "^1.6.5",
+    "passport": "^0.4.1",
+    "passport-local": "^1.0.0",
+    "path": "^0.12.7",
+    "sequelize": "^5.21.3"
   }
 }
